Move Application unit tests to its own module
- #22 moved Application data class object to its own module, but
  unit tests were not moved and were together with the analyze.
